fix(bedrock): applying changes per feedback
